modules/module-2a/requirements.txt
fastapi
uvicorn
mysql-connector-python
pika
pydantic